Consolidate on src/decorators/base.ts
