+++ packages/reactive-element/src/legacy-decorators/query-assigned-nodes.ts
@@ -11,7 +11,7 @@
  * not an arrow function.
  */
 import type {ReactiveElement} from '../reactive-element.js';
-import type {Interface} from './base.js';
+import type {Interface} from '../decorators/base.js';
 import {queryAssignedNodes as standardQueryAssignedNodes} from '../std-decorators/query-assigned-nodes.js';
 
 /**
